+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/SubjectCredentialStore.kt
@@ -261,7 +261,6 @@ interface SubjectCredentialStore {
         fun getDcApiId(): String = when (this) {
             is Vc -> vc.jwtId
             is SdJwt -> sdJwt.jwtId
-                ?: sdJwt.subject
                 ?: joseCompliantSerializer.encodeToString(sdJwt).toByteArray().sha256().toHexString()
 
             is Iso -> coseCompliantSerializer.encodeToByteArray(issuerSigned).sha256().toHexString()